refactor(lint): update import paths with simple_directive_paths rule
Adjusted import paths in multiple widget files to ensure consistency
and improve readability.
