+++ packages/world_countries/example/lib/ui/main/widgets/menu_button.dart
@@ -3,11 +3,11 @@ import "dart:async" show unawaited;
 import "package:flutter/material.dart";
 import "package:world_countries/helpers.dart";
 
-import "../../../../../routing/route_state.dart";
-import "../../../../helpers/extensions/parsed_route_settings_extension.dart";
+import "../../../helpers/extensions/parsed_route_settings_extension.dart";
 import "../../../helpers/extensions/string_extension.dart";
 import "../../../model/constants.dart";
 import "../../../notifications/notifications_center.dart";
+import "../../../routing/route_state.dart";
 import "../dialogs/about_app_dialog.dart";
 
 class MenuButton extends StatelessWidget {
