+++ packages/world_countries/example/lib/ui/settings/widgets/sliders/shadow_opacity_slider.dart
@@ -1,6 +1,6 @@
 import "package:flutter/material.dart";
 
-import "../../../../../helpers/extensions/shadows_flag_theme_controller_extension.dart";
+import "../../../../helpers/extensions/shadows_flag_theme_controller_extension.dart";
 import "settings_tile_slider.dart";
 
 final class ShadowOpacitySlider extends SettingsTileSlider {
