+++ packages/world_countries/lib/src/widgets/generic_widgets/iso_tile.dart
@@ -5,7 +5,7 @@ import "package:world_flags/world_flags.dart";
 
 import "../../extensions/models/iso_semantics_extension.dart";
 import "../../model/item_properties.dart";
-import "../generic_widgets/list_item_tile.dart";
+import "list_item_tile.dart";
 
 /// A tile widget that displays information about the ISO object.
 class IsoTile<T extends IsoTranslated> extends ListItemTile<T> {
